my-app/src/server/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
boto3>=1.26.0
botocore>=1.29.0
python-jose>=3.3.0
requests>=2.28.0
python-dotenv>=1.0.0
agentmail>=1.0.0